Implement validation (#88)
- Also removed some alternative datetime format tests as customizing the
time formats is too frustrating
- Added k6 tests to CI as they now pass

+++ go-budget-api/internal/entities/transaction.go
@@ -11,7 +11,7 @@ type Transaction struct {
 	Id           string     `gorm:"type:uuid;primary_key"`
 	Category     string     `gorm:"type:varchar(50);NOT NULL"`
 	Description  string     `gorm:"type:varchar(200);NOT NULL"`
-	Value        float32    `gorm:"type:decimal;NOT NULL"`
+	Value        float64    `gorm:"type:decimal;NOT NULL"`
 	Timestamp    time.Time  `gorm:"type:timestamp(3);NOT NULL"`
 	Credential   Credential `gorm:"constraint:OnUpdate:CASCADE,OnDelete:CASCADE"`
 	CredentialId string     `gorm:"NOT NULL"`
