+++ go-budget-api/internal/models/budget_report.go
@@ -1,9 +1,9 @@
 package models
 
 type BudgetReport struct {
-	TransactionsSum   float32 `json:"transactions_sum"`
-	ExpensesSum       float32 `json:"expenses_sum"`
-	IncomesSum        float32 `json:"incomes_sum"`
+	TransactionsSum   float64 `json:"transactions_sum"`
+	ExpensesSum       float64 `json:"expenses_sum"`
+	IncomesSum        float64 `json:"incomes_sum"`
 	TransactionsCount int32   `json:"transactions_count"`
 	ExpensesCount     int32   `json:"expenses_count"`
 	IncomesCount      int32   `json:"incomes_count"`
